Safari-609.1.11.1/LayoutTests/js/function-constructor-error-expected.txt

This test checks that the Function constructor detects some syntax errors correctly (bug#59795).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Function('(i + (j)') threw exception SyntaxError: Unexpected token '}'. Expected ')' to end a compound expression..
PASS Function('return (i + (j)') threw exception SyntaxError: Unexpected token '}'. Expected ')' to end a compound expression..
PASS successfullyParsed is true

TEST COMPLETE
